requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.3.2
crispy-bootstrap5==2024.2
dj-database-url==2.2.0
Django==4.2.1
django-cors-headers==4.4.0
django-crispy-forms==2.2
django-environ==0.11.2
django-filter==24.3
django-select-multiple-field==0.4.2
djangorestframework==3.15.2
Faker==30.1.0
idna==3.10
Markdown==3.7
pillow==10.4.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
requests==2.32.3
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
